partial bnf, do it in half min

+++ L1/ProblemeImplementare/cerc.cpp
@@ -4,10 +4,10 @@
 int main()
 {
     std::cout << "R =";
-    double raza;
+    float raza;
     std::cin >> raza;
-    const double a = 3.14 * raza * raza;
-    const double p = 2 * 3.14 * raza;
+    const float a = 3.14 * raza * raza;
+    const float p = 2 * 3.14 * raza;
     std::cout << "P = " << p << std::setprecision(2) << "\n";
     std::cout << "A =" << a << std::setprecision(2) << "\n";
     return 0;
